mads

+++ A4/libs/num_int1/include/num_int1.h
@@ -4,6 +4,6 @@
 #include <math.h>
 #include <stdio.h>
 
-double num_int1(double (*f)(double), double lim1, double lim2, int delta); 
+double num_int1(double (*f)(double), double lim1, double lim2, double delta); 
 
 #endif
